+++ src/TodoForm.jsx
@@ -1,12 +1,20 @@
 import React, { useEffect, useState } from 'react';
 
 import AddInput from './AddInput';
+import TodoTemplete from './TodoTemplete';
 
-function TodoForm({ handleAddTodo, templete }) {
+const templetes = ['1시간 홈트', '1일 1 야채', '1일 1 과일'];
+
+function TodoForm({ handleAddTodo }) {
   const [value, setValue] = useState('');
+  const [selectedTemplete, setSelectedTemplete] = useState(null);
 
   function handleInput(event) {
     setValue(event.target.value);
+  }
+
+  function handleSelectTemplete(item) {
+    setSelectedTemplete(item);
   }
 
   function clearInput() {
@@ -19,16 +27,20 @@ function TodoForm({ handleAddTodo, templete }) {
     handleAddTodo(value);
 
     clearInput();
+    setSelectedTemplete(null);
   }
 
   useEffect(() => {
-    setValue(templete || '');
-  }, [templete]);
+    setValue(selectedTemplete || '');
+  }, [selectedTemplete]);
 
   return (
-    <form onSubmit={onSubmit}>
-      <AddInput value={templete || value} handleInput={handleInput} />
-    </form>
+    <>
+      <TodoTemplete templetes={templetes} onClickTemplete={handleSelectTemplete} />
+      <form onSubmit={onSubmit}>
+        <AddInput value={selectedTemplete || value} handleInput={handleInput} />
+      </form>
+    </>
   );
 }
